+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [2.0.0] - 2025-02-08
 
 ### Added
 
@@ -36,6 +36,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Initial release.
 
-[Unreleased]: https://github.com/newAM/oidc_pages/compare/v1.1.0...HEAD
+[Unreleased]: https://github.com/newAM/oidc_pages/compare/v2.0.0...HEAD
+[2.0.0]: https://github.com/newAM/oidc_pages/compare/v1.1.0...v2.0.0
 [1.1.0]: https://github.com/newAM/oidc_pages/compare/v1.0.0...v1.1.0
 [1.0.0]: https://github.com/newAM/oidc_pages/releases/tag/v1.0.0